Create mesh subfolder and overload mesh functions
